requirements.txt
peewee~=3.17.5
requests~=2.32.3
python-telegram-bot==21.3
pytz==2024.1
python-dotenv~=1.0.1